vid_resolutions[mode].address now uses real2ptr from modeinfo.pptr.  No need to use the old code anymore!  Tested for about 30 minutes on q2.playground.ru.  Seems OK, but ifdef in place incase its no good.

+++ client/cl_scrn.c
@@ -569,14 +569,14 @@ void SCR_RunConsole (void)
 	
 	if (scr_conlines < scr_con_current)
 	{
-		scr_con_current -= scr_conspeed->value*cls.frametime;
+		scr_con_current -= scr_conspeed->value*(float)cls.frametime;
 		if (scr_conlines > scr_con_current)
 			scr_con_current = scr_conlines;
 
 	}
 	else if (scr_conlines > scr_con_current)
 	{
-		scr_con_current += scr_conspeed->value*cls.frametime;
+		scr_con_current += scr_conspeed->value*(float)cls.frametime;
 		if (scr_conlines < scr_con_current)
 			scr_con_current = scr_conlines;
 	}
